refactor(resolume): split driver.rs tests + tighten field visibility + add Err-path test + dashboard cleanup
Addresses review weaknesses on PR #54 in one focused commit:

1. driver.rs split: extract `mod tests` to driver_tests.rs via #[path].
   driver.rs: 999 -> 512 lines (relieves the 999/1000 cap pressure
   that bit us on PR #53 and would block any future driver edit).

2. Tighten HostDriver field visibility: last_refresh_ok, last_refresh_ts,
   recovery_tx, health_tx are private. consecutive_failures and
   circuit_breaker_open stay accessible to descendant test module via
   private-to-children rule. clip_mapping stays pub(crate) for
   handlers.rs sibling access.

3. cached_position_ms doc: spell out the ~500ms staleness bound on the
   recovery re-emit (one Position-tick worst case).

4. New test: started_event_with_malformed_lyrics_warns_and_clears_state
   (covers the Err(e) branch of load_lyrics_for_video — was previously
   uncovered, only Ok(None) had a regression test).

5. ResolumeHealthCard cleanup: wire on_cleanup to a cancellation flag
   so the spawn_local poll loop exits when the component unmounts.
   Previously the loop ran for the lifetime of the wasm runtime.
